Day 1
I think I might have to do project 3. There is a portion of code in this that is not working the way it should and it is driving me crazy. I got the first part of this done and now I have to figure out how to do multi-nested for loops to write some HTML code of the sorted list.